--- a/typed-racket-lib/typed-racket/base-env/base-types-extra.rkt
+++ b/typed-racket-lib/typed-racket/base-env/base-types-extra.rkt
@@ -10,11 +10,21 @@
               ...
               (provide nm) ...)]))
 
+(define-syntax (define-other-props stx)
+  (syntax-case stx ()
+    [(_ nm ...)
+     #'(begin (define-syntax nm prop-name-error)
+              ...
+              (provide nm) ...)]))
+
 ;; special type names that are not bound to particular types
 (define-other-types
   -> ->* case-> U Union ∩ Intersection Rec All Opaque Vector
   Parameterof List List* Class Object Row Unit Values AnyValues Instance Refinement
-  pred Struct Struct-Type Prefab Top Bot Distinction Sequenceof)
+  pred Struct Struct-Type Prefab Distinction Sequenceof Refine)
+
+(define-other-props
+  Top Bot !)
 
 (provide (rename-out [All ∀]
                      [U Un]