--- a/typed-racket-lib/typed-racket/rep/core-rep.rkt
+++ b/typed-racket-lib/typed-racket/rep/core-rep.rkt
@@ -29,7 +29,7 @@
          def-values
          def-prop
          def-object
-         def-pathelem)
+         def-path-elem)
 
 
 ;;************************************************************
@@ -136,7 +136,7 @@
 ;;--------------
 
 ;; e.g. car, cdr, etc
-(def-rep-class PathElem #:printer print-pathelem #:define-form def-pathelem)
+(def-rep-class PathElem #:printer print-pathelem #:define-form def-path-elem)
 
 
 ;;----------