# In this PR
This PR adds some of the basic atoms for refinement types, namely the following:

- Linear Expression objects (see object-rep.rkt)

- Linear Inequality propositions (see prop-rep.rkt)

- a solver for systems of linear inequalities (see logic/ineq.rkt)

- intersection types now carry any logical refinement proposition (tt when there is none) -- I chose to use intersection types for this role to avoid duplicating code. There are places in TR where, for example, we currently match on the literal type `(Vectorof τ)`, and these parts of the type checker should also work with refinements, intersections, etc. Now there just needs to be 1 additional line added (for intersections) instead of 2 or more (for intersections, refinements, etc).

- user-level syntax for the above abstractions

- tests for the above abstractions

# Not in this PR

- functions w/ dependent arguments

- the necessary tweaks for type inference w/ refinements to work on real programs (namely plumbing objects through the constraint generation process)

- actually producing refinement types by default anywhere when typechecking

# Other misc functionality

- the types of symbolic objects (and not just identifiers) can now also stored in the lexical environment (e.g. if an identifier is an alias for a linear expression, we have to keep the type of the linear expression that it aliases around so occurrence typing of more specific integer types still works for the identifier -- see type-env-structs.rkt, lexical-env.rkt, etc)

- `combine-props` in tc-metafunctions.rkt works harder to prove things (needed for refinement subtyping)

- there is now two contradiction functions: `atomic-contradiction?` and `contradiction?`. the lighter weight one is used for normalizing propositions, the stronger one is used for `combine-props` (i.e. when we're actually trying to prove/derive things)

- subtyping now considers refinement propositions and calls the logic portion of TR when necessary

- printing can print refinement types (printer.rkt)

- the parser (private/parse-type.rkt) has a new, compositional grammar for logical propositions. This is currently only used for parsing the `ψ` in `(Refine [x : τ] ψ)` but will also be used later for parsing function types.

- There is a new low-level mechanism for instantiating/abstracting term-level bindings in Reps (i.e. its like tc-subst but doesn't reason about the semantics of the types w.r.t. subtyping, overlap, etc -- see type-rep.rkt `subst-names`)

# To Do (before possibly ready for merge)

- [x] performance impact?
- [x] subtyping shouldn't cache for refinements
- [x] more tests (subtyping (logic, w/ envs), substitution, more?)
- [x] document syntax for propositions