@samth 2017-03-15
Initial thoughts (before reading the code):

1. What are the performance impacts?
2. Do the things not in the PR impact our confidence in the correctness/testing of the things here?
3. What's the syntax, and is it documented?