@pnwamk 2017-03-15
3. Syntax: The new syntax is not documented in this PR -- it is currently as follows:

Refinements are written `(Refine [x : τ] p)` where `τ` is valid type syntax (unchanged) and `p` is defined by:
```
o ::= x | (car o) | (cdr o)
t ::= o | (* n o)
l ::= n | t | (+ t t ...)
p ::= Top | Bot | (: l t) | (! l t) | (and p p ...) | (or p p ...) | (not p) 
          | (when p p) | (unless p p) | (if p p p) 
          | (<= l l) | (< l l) | (> l l) | (>= l l) | (= l l)
```

2. Correctness: This PR should be mostly testable and self contained (although it does need a little more testing at the moment). The internal data structures that are added area straightforwardly testable, and the tweaks to subtyping and the logic are as well.

1. Performance is largely unaffected (which was sort of what I expected since the new code isn't hit in general by non-refinement types, etc), with a minor caveat -- most of the micro benchmarks I use for TR were unaffected, except:

`math-flonum.rkt` got slightly slower (3.45 to 3.7sec) -- this seems to primarily stem from the fact that our `normalize-id` function (which maps free-identifier=? ids to the same id) now uses a weak hash table instead of a standard free-id-table. This is necessary now since we gensym identifiers for refinements (and later we will for dependent function checking).

`schml-interp-cast.rkt` got slightly faster (23seconds to 21 seconds by my measurement)... no idea why (might be that the stronger logic used for extending the environment more quickly rules some cases out as uninhabitable... but I'm not sure)