@pnwamk 2017-03-15
e076284 doesn't seem to affect typechecking type either