@pnwamk 2017-03-17
I've added additional testing.

Unless there's things we want to change, the only other possible addition would be to add documentation for Refinement types, unless we want to add that with updated function syntax (which might be a more natural place to put the grammar for propositions... I dunno)