@pnwamk 2017-03-17
Oh, two more things I need to add to this PR:
- [ ] well-formedness of the refinements (e.g. lexp expect integers, car/cdr expect pairs, vector-length expects vectors, etc)
- [ ] contract generation!