@bennn 2017-03-22
> What are the performance impacts?

I compiled the Typed Racket programs from the [gradual-typing-performance](https://github.com/nuprl/gradual-typing-performance) repo on Racket v6.8 and this branch.
- compile times are 0.5 to 3 seconds slower
- runtimes are almost the same
  - but for some reason `fsmoo` and `take5` (object oriented programs) are consistently 50 milliseconds slower

Here are the raw timings (compile times first, scroll down for running times)
https://docs.google.com/spreadsheets/d/1S9PccVnGmzh8y0GhWbsjU6-JQuPoZ_6dTQ9nTw21stk/edit?usp=sharing

Here are the programs. If you unpack this archive, edit the Makefile to point to your "control" and "experimental" Racket binaries, and run `make` you should be able to reproduce this little experiment.
[refine.tar.gz](https://github.com/racket/typed-racket/files/860321/refine.tar.gz)