@mfelleisen 2017-03-22
@benn Thanks for keeping an eye open on performance. In my capacity as PLT’s CPS, I herewith appoint you to ‘performance czar’ of Typed Racket. 




> On Mar 22, 2017, at 12:26 AM, Ben Greenman <notifications@github.com> wrote:
> 
> What are the performance impacts?
> 
> I compiled the Typed Racket programs from the gradual-typing-performance <https://github.com/nuprl/gradual-typing-performance> repo on Racket v6.8 and this branch.
> 
> compile times are 0.5 to 3 seconds slower
> runtimes are almost the same
> but for some reason fsmoo and take5 (object oriented programs) are consistently 50 milliseconds slower
> Here are the raw timings (compile times first, scroll down for running times)
> https://docs.google.com/spreadsheets/d/1S9PccVnGmzh8y0GhWbsjU6-JQuPoZ_6dTQ9nTw21stk/edit?usp=sharing <https://docs.google.com/spreadsheets/d/1S9PccVnGmzh8y0GhWbsjU6-JQuPoZ_6dTQ9nTw21stk/edit?usp=sharing>
> Here are the programs. If you unpack this archive, edit the Makefile to point to your "control" and "experimental" Racket binaries, and run make you should be able to reproduce this little experiment.
> refine.tar.gz <https://github.com/racket/typed-racket/files/860321/refine.tar.gz>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub <https://github.com/racket/typed-racket/pull/510#issuecomment-288295385>, or mute the thread <https://github.com/notifications/unsubscribe-auth/ABf5-9vx0cAOCx7lTgqY97P-8WF9U1TUks5roKLzgaJpZM4Md5ry>.
>