@pnwamk 2017-03-27
This PR is pretty much wrapped up. All the intended features are in. More vanilla programs can typecheck (e.g. we do a more thorough job proving expected propositions now) and there is a flag for typechecking with linear integer arithmetic reasoning.

Here are the current performance impacts (based on a small sample of files/programs):

![image](https://cloud.githubusercontent.com/assets/3662058/24363309/77f60d92-12dd-11e7-8664-d724bd87480a.png)

Things are a little slower compared to HEAD, but we're still doing well compared to TR before some of the recent performance work that went into the 6.8 release. Hopefully we can identify and further mitigate any performance overhead that is added here, but for the moment we think this is reasonable for this PR.